pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "slidegeist"
version = "2025.10.22"
description = "Extract slides and transcripts from lecture videos with minimal dependencies"
readme = "README.md"
license = { text = "MIT" }
requires-python = ">=3.10"
authors = [
    { name = "Christopher Albert" }
]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Education",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: Scientific/Engineering",
]
dependencies = [
    "faster-whisper>=1.0.0",
    "numpy>=1.24.0",
    "opencv-python>=4.8.0",
    "yt-dlp>=2024.0.0",
    "torch>=2.3.0",
    "transformers>=4.44.0",
    "accelerate>=0.30.0",
    "safetensors>=0.4.3",
    "qwen-vl-utils>=0.0.14",
    "pytesseract>=0.3.10",
    "Pillow>=10.0.0",
    "sentencepiece>=0.1.99",
]

[project.optional-dependencies]
mlx = [
    "mlx-whisper>=0.1.0",
    "mlx-vlm>=0.3.4",
]
cuda = [
    "bitsandbytes>=0.43.0",
]
dev = [
    "pytest>=7.0",
    "pytest-cov>=4.0",
    "ruff>=0.1.0",
    "mypy>=1.0",
]

[project.scripts]
slidegeist = "slidegeist.cli:main"

[tool.ruff]
line-length = 100
target-version = "py310"

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
ignore = ["E501"]  # Line too long (handled by formatter)

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.pytest.ini_options]
testpaths = ["tests"]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
addopts = "--cov=slidegeist --cov-report=term-missing -m 'not manual'"
markers = [
    "slow: marks tests as slow (deselect with '-m \"not slow\"')",
    "manual: marks tests requiring manual intervention or validation",
]
